chore: ConnectionStatus Enum 분리
