+++ HonestHouse/HonestHouse/Sources/Presentation/CameraConnection/Shared/ConnectionStatus.swift
@@ -1,0 +1,15 @@
+//
+//  ConnectionStatus.swift
+//  HonestHouse
+//
+//  Created by Rama on 11/3/25.
+//
+
+import Foundation
+
+enum ConnectionState: Equatable {
+    case disconnected
+    case connecting
+    case connected
+    case failed(String)
+}
